refactor(backend): decouple JWT validation from Auth domain layer
- Created `TokenVerificationProtocol` interface in `domain/auth/interfaces.py`.
- Moved `PyJWKClient` and direct `jwt` framework usage out of `AuthService`.
- Created new `SupabaseTokenVerifier` in `infrastructure/external/jwt.py` implementing the interface.
- Refactored `AuthService` to accept the `TokenVerificationProtocol` via dependency injection.
- Updated FastAPI dependency providers in `dependencies.py` to inject `SupabaseTokenVerifier`.
- Updated WebSocket manual `AuthService` instantiation in `websocket.py`.
- Updated test cases in `test_auth.py` to support the injected verifier dependency.
